# api/openapi.yaml
openapi: 3.0.0
info:
  title: Meeting Management System
  version: 1.1.0
paths:
  /meetings:
    post:
      summary: Create a new meeting
      parameters:
        - in: query
          name: date_time
          schema:
            type: string
            format: date-time
          required: true
        - in: query
          name: meet_link
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meet_id:
                    type: string

  /users:
    post:
      summary: Register a new user
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
          required: true
        - in: query
          name: org
          schema:
            type: string
          required: true
        - in: query
          name: meet_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string

  /meetings/{meet_id}:
    get:
      summary: Get a meeting by meet_id
      parameters:
        - in: path
          name: meet_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Meeting retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meet_id:
                    type: string
                  meet_link:
                    type: string
                  date_time:
                    type: string
                    format: date-time


  /answers:
    put:
      summary: Submit answers to questions
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
          required: true
        - in: query
          name: name
          schema:
            type: string
          required: true
        - in: query
          name: answer_1
          schema:
            type: string
          required: true
        - in: query
          name: answer_2
          schema:
            type: string
          required: true
        - in: query
          name: answer_ai
          schema:
            type: string
          required: true
        - in: query
          name: meet_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Answers submitted successfully

  /answers/{meet_id}:
    get:
      summary: Get all answers for a meeting
      parameters:
        - in: path
          name: meet_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of answers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    answer_1:
                      type: string
                    answer_2:
                      type: string
                    answer_ai:
                      type: string
  /questions:
    post:
      summary: Add a question
      parameters:
        - in: query
          name: meet_id
          schema:
            type: string
          required: true
        - in: query
          name: question
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Question added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meet_id:
                    type: string
                  question:
                    type: string
  /questions/{meet_id}:
    get:
      summary: Get all questions for a meeting
      parameters:
        - in: path
          name: meet_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    meet_id:
                      type: string
                    question: